perf(site): lazy-load the hero editor behind the boot loader
The landing page's only eager script is now a ~114 byte loader. Everything
heavy — the CodeMirror mini-editor, the player bridge, and now React — moved
to src/hero-app.tsx and arrives through a dynamic import, so the first paint
depends on the static shell alone. The boot loader (static markup, added in
#510) covers the gap and is dismissed by the same player-readiness signal as
before.

The hero's one piece of chrome, the status dot, becomes a React island over
the shared `Store`, completing the site's React migration. The `__hero` seam
keeps its exact shape and now reads the store directly, so it never waits on
a render.

esbuild builds the hero in its own pass with `splitting: true` — the only way
it emits the dynamic import as a separate file — rather than turning splitting
on for every entry, which would rewrite sandbox/ide/demo-editor into multi-file
loads for no benefit.
